refactor(test): 투입 금액 입력에 대한 테스트 분리
