+++ src/test/java/lotto/ApplicationTest.java
@@ -54,46 +54,6 @@ class ApplicationTest extends NsTest {
         });
     }
 
-    @Test
-    void 예외_테스트_빈칸_입력(){
-        assertSimpleTest(() -> {
-            runException(" ");
-            assertThat(output()).contains(ERROR_MESSAGE);
-        });
-    }
-
-    @Test
-    void 예외_테스트_문자_입력(){
-        assertSimpleTest(() -> {
-            runException("$10000");
-            assertThat(output()).contains(ERROR_MESSAGE);
-        });
-    }
-
-    @Test
-    void 예외_테스트_음수_입력(){
-        assertSimpleTest(() -> {
-            runException("-1000");
-            assertThat(output()).contains(ERROR_MESSAGE);
-        });
-    }
-
-    @Test
-    void 예외_테스트_최소_금액_미만_입력(){
-        assertSimpleTest(() -> {
-            runException("999");
-            assertThat(output()).contains(ERROR_MESSAGE);
-        });
-    }
-
-    @Test
-    void 예외_테스트_거스름돈_존재(){
-        assertSimpleTest(() -> {
-            runException("1004");
-            assertThat(output()).contains(ERROR_MESSAGE);
-        });
-    }
-
     @Override
     public void runMain() {
         Application.main(new String[]{});
